move Cucumber::MultilineArgument::DataTable::DiffMatrices to its own file.
